feat(mixins): richable MathMixins

+++ i2f-jdk/i2f-math/src/main/java/i2f/math/MathUtil.java
@@ -656,7 +656,7 @@ public class MathUtil {
      * @return
      */
     public static long fibonacci(int i) {
-        return Fibonacci.get(i);
+        return Fibonacci.get(i).longValue();
     }
 
     /**
